+++ geowatch/cli/smartflow/run_sv_datagen.py
@@ -78,6 +78,16 @@ class SVDatasetConfig(scfg.DataConfig):
             config for SV_Cropping.
             '''))
 
+    input_region_models_asset_name = scfg.Value('cropped_region_models_bas', type=str, required=False, help=ub.paragraph(
+            '''
+            Which region model assets to use as input
+            '''))
+
+    input_site_models_asset_name = scfg.Value('cropped_site_models_bas', type=str, required=False, help=ub.paragraph(
+            '''
+            Which site model assets to to use as input
+            '''))
+
 
 def main():
     config = SVDatasetConfig.cli(strict=True)
@@ -124,8 +134,9 @@ def run_generate_sv_cropped_kwcoco(config):
     ingressed_assets = smartflow_ingress(
         input_path,
         ['kwcoco_for_sc',
-         'cropped_region_models_bas',
-         'cropped_site_models_bas'],
+         config.input_region_models_asset_name,
+         config.input_site_models_asset_name,
+         ],
         ingress_dir,
         aws_profile,
         dryrun)
@@ -145,7 +156,7 @@ def run_generate_sv_cropped_kwcoco(config):
         raise RuntimeError("Couldn't parse 'region_id' from input region file")
 
     # Paths to inputs generated in previous pipeline steps
-    bas_region_path = ub.Path(ingressed_assets['cropped_region_models_bas']) / f'{region_id}.geojson'
+    bas_region_path = ub.Path(ingressed_assets[config.input_region_models_asset_name]) / f'{region_id}.geojson'
     ta1_sc_kwcoco_path = ingressed_assets['kwcoco_for_sc']
 
     node_state.print_current_state(ingress_dir)